Created a bas class "Saveable" which LawWorld Prison and PrisonCell inherit from.
This provides an overridable save method for each class which only
writes to file the values which have changed, or can be flagged to
perform a clean and write, which will delete the node in the config
file and write it fresh.

Since saveable classes are loaded from file on plugin startup that are
either "active" or not. When active changes are marked to be written to
file, this aids the setup plugin onEnable method to build instances
without writing back to the config file.

Also instances are saved in a cascading fashion, from LawWorld to
Prison, down to PrisonCell. Each container will call save on all child
references and each child determines what it needs to save (if
anything).

The result of this is that the "/law imprison" and "/law free" commands
both fully work, and the "/law prison" command supports "create",
"delete", "list", "select", "setdefaultcell", "setexitpoint" commands,
but currently not "addcell" and "removecell".