Refactored commands such that each has its own class.
OK, this is a monster commit for which I should have created a branch.
Maybe I am STILL thinking in SVN mode, or it could be that I knew it
would work, or that I wouldn't release the plugin without making this
change. Regardless the change was a success and the plugin is the
better for it.