Added the confiscation of imprisoned players inventories (items and armor), which are returned after being freed.
Inventories are freed and saved in-memory, but are also saved to file.
If the server is reloaded or crashes the inventories will no longer be
in memory and will be loaded from file instead.

If a player is offline when an inventory change takes effect
(imprisoning or freeing) then the the confiscation or return of
inventory items will be taken care of on rejoining the server.